Refactor filter card
